requirements.txt
apache-airflow==2.6.0
apache-airflow-providers-google==8.10.0
google-cloud-bigquery==3.9.0
google-cloud-storage==2.8.0
PyYAML==6.0
pandas==1.5.3
google-auth==2.16.0
pyarrow==12.0.0
requests==2.28.2
